收藏夹修改，新增迁移/路由/命令

# 收藏夹修改需求：
## 迁移：
  bookmark表新增字段 default,值1/0.类型boolean,默认0,不为空。

## 命令：
 - 一条artisan 命令,为所有用户生成一个default值为1,名称为default的记录。已有就修改default值。  
 - 一条artisan 命令,检查所有用户,列出有两条或以上default值为1的用户。  

## 后端修改：
 - default字段值为1的收藏夹不允许被修改&删除。  
 - 检查创建收藏夹代码,不允许传入default值,由系统提供。  
 - 用户创建时,创建默认收藏夹代码需要修改default值为1。
 - 加个api bookmark/default,返回默认收藏夹。


## model层修改：
 - 增加一个函数canModify方法判断当前是否可修改数据，判断依据为：当前记录的default值不为1，则可以修改。

## 自测点：
排序  | 测试类型 | 测试要点| 测试内容  |  测试结果  | 备注
---|---|---|---|---|---
1|迁移|执行迁移,查看default字段是否正常增加删除|执行迁移,bookmarks表增加default字段,值为0.执行迁移回滚,bookmarks表的default字段消失|PASS
2-1|前端测试|新注册用户的收藏夹值为1(不能修改)|注册一个用户,然后在收藏夹页面尝试修改名称,不能修改|PASS
2-2|前端测试|默认收藏夹不能修改/删除|在Bookmark界面尝试编辑default收藏夹名称和尝试删除,不能编辑和删除|PASS|返回值待定
2-3|前端测试|用户自行增加的收藏夹default值为0|在Bookmark页面新增一个收藏夹,检查数据库,该收藏夹的default值为0|PASS
2-4|前端测试|同一用户名下不能有2个同名文件夹，不同用户可以|登录账号A，创建一个收藏夹A，再创建一个收藏夹B，然后把B的名称修改为A，保存，提示收藏夹名称已存在；然后登录账号B，创建收藏夹A，成功创建。|PASS
3|新增api测试|返回默认收藏夹|使用Postman模拟get请求域名/bookmark/default,返回值为1个对象类似于：{"id":1039,"uid":57556,"name":"default","default":1,"item_count":0},default值为1|PASS
4-1|命令测试|修复所有用户的默认收藏夹|执行命令php artisan bba:checkBookmark --fix-all,被修复的用户将列表显示,修复后用户名下的收藏夹中有1个名称为default,default值为1|PASS
4-2|命令测试|检出有2个及以上默认收藏夹的用户|执行命令php artisan bba:checkBookmark --check-all,被检出的用户将列表显示|PASS|是否需要将检出的结果以类似日志文件一样存一份
4-3|命令测试|包含以上2个检查|执行命令php artisan bba:checkBookmark,修复用户的默认收藏夹,检出收藏夹异常用户,执行结果列表显示|PASS|为达到效果可以把上面命令涉及到的记录还原再执行，执行结果包含上面2个结果

See merge request !262

+++ app/Http/Controllers/BookmarkController.php
@@ -19,7 +19,7 @@ class BookmarkController extends Controller
      * 显示收藏目录
      */
     public function index(Request $req)
-    {   
+    {
         //显示收藏夹之前需要检查是否为合法用户
         if (intval($req->uid) != Auth::user()->id) {
             return $this->responseError("No Permission", -1);
@@ -44,10 +44,10 @@ class BookmarkController extends Controller
             按权限限制收藏夹目录数量及检测有效性
         */
         $bookmark = new Bookmark();
-        $bookmarkListUsage = Auth::user()->getUsage('bookmark_list');
-        $bookmarkLists = Bookmark::where("uid", Auth::user()->id)->get();
-        $bookmarkListCount = Bookmark::where("name", $request->name)
-                                       ->where("uid", Auth::user()->id)
+        $user = Auth::user();
+        $bookmarkListUsage = $user->getUsage('bookmark_list');
+        $bookmarkLists = Bookmark::where("uid", $user->id)->get();
+        $bookmarkListCount = $user->bookmarks()->where("name", $request->name)
                                        ->count();
         if (count($bookmarkLists) >= $bookmarkListUsage[1]) {
             return $this->responseError(trans('messages.bookmark_num_limit'), -4498);
@@ -56,12 +56,16 @@ class BookmarkController extends Controller
             return $this->responseError(trans('messages.bookmark_existed'), -4496);
         }
         $validator = Validator::make($request->all(), ['name' => 'required|between:1,25']);
-        if ($validator->fails()) 
+        if ($validator->fails())
         {
             return $this->responseError(trans('messages.bookmark_length_limit'), -4497);
         }
-        $bookmark->uid = Auth::user()->id;
+        if (!$bookmark->canModify()) {
+            return $this->responseError(trans('messages.bookmark_can_not_modify'), -4500);
+        }
+        $bookmark->uid = $user->id;
         $bookmark->name = $request->name;
+        $bookmark->default = 0; // 用户自行加入的
         $bookmark->save();
         return $bookmark;
     }
@@ -79,7 +83,8 @@ class BookmarkController extends Controller
     public function update(Request $req, $id)
     {
         //编辑收藏夹之前需要检查是否为合法用户
-        if (intval($req->uid) != Auth::user()->id) {
+        $user = Auth::user();
+        if (intval($req->uid) != $user->id) {
             return response(["code"=>-1, "desc"=>"No Permission"], 422);
         }
         $bookmark = Bookmark::where("id", $id)->first();
@@ -89,17 +94,24 @@ class BookmarkController extends Controller
         /*if (!Auth::user()->can('update', $bookmark)) {
             return response(["code"=>-1, "desc"=>"No Permission"], 501);
         }*/
-        //限制目录名称长度
+        if (!$bookmark->canModify()) {
+            return $this->responseError(trans('messages.bookmark_can_not_modify'), -4500);
+        }
+        // 同一用户名下不允许有2个同名收藏夹
+        if ($user->bookmarks()->where("name", $req->name)->where('id', '<>', $req->id)->first()) {
+            return $this->responseError(trans('messages.bookmark_existed'), -4496);
+        }
+        // 限制目录名称长度
         $validator = Validator::make($req->all(), ['name' => 'required|between:1,25']);
-        if ($validator->fails()) 
+        if ($validator->fails())
         {
             return $this->responseError(trans('messages.bookmark_length_limit'), -4497);
         }
         $update = false;
         foreach($req->all() as $key=>$value) {
             if (isset($bookmark[$key]) && $value != $bookmark->$key) {
-               $bookmark->$key = $value;
-               $update = true;
+                $bookmark->$key = $value;
+                $update = true;
             }
         }
         $bookmark->save();
@@ -123,11 +135,23 @@ class BookmarkController extends Controller
         /*if (!Auth::user()->can('delete', $bookmark)) {
             return response(["code"=>-1, "desc"=>"No Permission"], 501);
         } */
+        if (!$bookmark->canModify()) {
+            return $this->responseError(trans('messages.bookmark_can_not_modify'), -4500);
+        }
         $items = $bookmark->items;
         foreach ($items as $item) {
             $item->delete();
         }
         $bookmark->delete();
         return ["code"=>0, "desc"=>"Success"];
+    }
+
+    public function getDefault()
+    {
+        if ($user = Auth::user()) {
+            return $user->bookmarks()->where('default', 1)->first();
+        } else {
+            return $this->responseError(trans('messages.bookmark_list_no_find'), -4495);
+        }
     }
 }
